+++ hudi-common/src/test/java/org/apache/hudi/common/testutils/HoodieTestUtils.java
@@ -113,8 +113,8 @@ public class HoodieTestUtils {
 
   public static HoodieTableMetaClient init(String basePath, HoodieTableType tableType, String bootstrapBasePath, HoodieFileFormat baseFileFormat) throws IOException {
     Properties props = new Properties();
-    props.setProperty(HoodieTableConfig.HOODIE_BOOTSTRAP_BASE_PATH_PROP.key(), bootstrapBasePath);
-    props.setProperty(HoodieTableConfig.HOODIE_BASE_FILE_FORMAT_PROP.key(),baseFileFormat.name());
+    props.setProperty(HoodieTableConfig.BOOTSTRAP_BASE_PATH.key(), bootstrapBasePath);
+    props.setProperty(HoodieTableConfig.BASE_FILE_FORMAT.key(), baseFileFormat.name());
     return init(getDefaultHadoopConf(), basePath, tableType, props);
   }
