+++ hudi-spark-datasource/hudi-spark/src/main/java/org/apache/hudi/bootstrap/SparkOrcBootstrapDataProvider.java
@@ -18,68 +18,21 @@
 
 package org.apache.hudi.bootstrap;
 
-import org.apache.hudi.DataSourceUtils;
-import org.apache.hudi.HoodieSparkUtils;
-import org.apache.hudi.avro.HoodieAvroUtils;
-import org.apache.hudi.avro.model.HoodieFileStatus;
-import org.apache.hudi.client.bootstrap.FullRecordBootstrapDataProvider;
 import org.apache.hudi.client.common.HoodieSparkEngineContext;
-import org.apache.hudi.common.bootstrap.FileStatusUtils;
 import org.apache.hudi.common.config.TypedProperties;
-import org.apache.hudi.common.model.HoodieRecord;
-import org.apache.hudi.common.util.Option;
-import org.apache.hudi.common.util.collection.Pair;
-import org.apache.hudi.exception.HoodieIOException;
-import org.apache.hudi.keygen.KeyGenerator;
-
-import org.apache.avro.generic.GenericRecord;
-import org.apache.hudi.keygen.factory.HoodieSparkKeyGeneratorFactory;
-import org.apache.spark.api.java.JavaRDD;
-import org.apache.spark.rdd.RDD;
-import org.apache.spark.sql.Dataset;
-import org.apache.spark.sql.SparkSession;
-
-import java.io.IOException;
-import java.util.List;
 
 /**
  * Spark Data frame based bootstrap input provider.
  */
-public class SparkOrcBootstrapDataProvider extends FullRecordBootstrapDataProvider<JavaRDD<HoodieRecord>> {
-  private final transient SparkSession sparkSession;
+public class SparkOrcBootstrapDataProvider extends SparkFullBootstrapDataProviderBase {
 
   public SparkOrcBootstrapDataProvider(TypedProperties props,
                                        HoodieSparkEngineContext context) {
     super(props, context);
-    this.sparkSession = SparkSession.builder().config(context.getJavaSparkContext().getConf()).getOrCreate();
   }
 
   @Override
-  public JavaRDD<HoodieRecord> generateInputRecords(String tableName, String sourceBasePath,
-                                                    List<Pair<String, List<HoodieFileStatus>>> partitionPathsWithFiles) {
-    String[] filePaths = partitionPathsWithFiles.stream().map(Pair::getValue)
-            .flatMap(f -> f.stream().map(fs -> FileStatusUtils.toPath(fs.getPath()).toString()))
-            .toArray(String[]::new);
-
-    Dataset inputDataset = sparkSession.read().orc(filePaths);
-    try {
-      KeyGenerator keyGenerator = HoodieSparkKeyGeneratorFactory.createKeyGenerator(props);
-      String structName = tableName + "_record";
-      String namespace = "hoodie." + tableName;
-      RDD<GenericRecord> genericRecords = HoodieSparkUtils.createRdd(inputDataset, structName, namespace, false,
-              Option.empty());
-      return genericRecords.toJavaRDD().map(gr -> {
-        String orderingVal = HoodieAvroUtils.getNestedFieldValAsString(
-                gr, props.getString("hoodie.datasource.write.precombine.field"), false);
-        try {
-          return DataSourceUtils.createHoodieRecord(gr, orderingVal, keyGenerator.getKey(gr),
-                  props.getString("hoodie.datasource.write.payload.class"));
-        } catch (IOException ioe) {
-          throw new HoodieIOException(ioe.getMessage(), ioe);
-        }
-      });
-    } catch (IOException ioe) {
-      throw new HoodieIOException(ioe.getMessage(), ioe);
-    }
+  protected String getFormat() {
+    return "orc";
   }
 }
